攻撃回数をIntからDoubleに変更
装備品の攻撃回数ボーナスが反映されないバグを修正。
10倍スケール方式を廃止し、Doubleで直接保持するように変更。

- ItemDefinition.CombatBonuses.attackCount: Int → Double
- CharacterValues.Combat.attackCount: Int → Double
- ItemMaster.jsonのattackCount値を1/10に変換
- SQLiteスキーマをREAL型に変更
- 装備ボーナス適用後に最低1回を保証

+++ EpikaTests/Battle/BattleTurnEngineTacticTests.swift
@@ -394,7 +394,7 @@ private enum BattleTestFactory {
                        hitRate: Int = 80,
                        evasion: Int = 5,
                        critical: Int = 0,
-                       attackCount: Int = 1,
+                       attackCount: Double = 1,
                        magicalHealing: Int = 0,
                        additionalDamage: Int = 0,
                        breathDamage: Int = 0) -> CharacterValues.Combat {
